requirements.txt
amqp==5.3.1 
annotated-types==0.7.0 
anyio==4.8.0 
billiard==4.2.1 
celery==5.4.0 
certifi==2025.1.31 
click-didyoumean==0.3.1 
click-plugins==1.1.1 
click-repl==0.3.0 
click==8.1.8 
colorama==0.4.6 
dnspython==2.7.0 
email-validator==2.2.0 
fastapi-cli==0.0.7 
fastapi==0.115.8 
h11==0.14.0 
httpcore==1.0.7 
httptools==0.6.4 
httpx==0.28.1 
idna==3.10 
jinja2==3.1.5 
kombu==5.4.2 
markdown-it-py==3.0.0 
markupsafe==3.0.2 
mdurl==0.1.2 
names==0.3.0 
prompt-toolkit==3.0.50 
pydantic-core==2.27.2 
pydantic==2.10.6 
pygments==2.19.1 
python-dateutil==2.9.0.post0 
python-dotenv==1.0.1 
python-multipart==0.0.20 
pyyaml==6.0.2 
redis==5.2.1 
rich-toolkit==0.13.2 
rich==13.9.4 
shellingham==1.5.4 
six==1.17.0 
sniffio==1.3.1 
starlette==0.45.3 
typer==0.15.1 
typing-extensions==4.12.2 
tzdata==2025.1 
uvicorn==0.34.0 
uvloop==0.21.0 
vine==5.1.0 
watchfiles==1.0.4 
wcwidth==0.2.13 
websockets==14.2 
